Revise MCPHosts code for better structure and error checking.
Creation of the hosts header is now done during the boostrap phase
instead of being tried anytime a process initialized MCPHosts.  This
allows for better error handling: we were previously ignoring
MCPFileCreateFile return code altogether.  There is also a better
separation of labour between MCPHostsShmemInit (initialize the necessary
structs in shared memory) and MCPHostsInit (attach to the already-
initialized shmem struct and read the file from disk if necessary).

We keep the whole thing in shared memory instead of having half of it in
private memory that needed to be separately palloc'ed in every forwarder
child process, and assign the lock IDs with LWLockAssign() instead of
having a fixed range of values; while this worked, it was error prone
(we had already introduced a bug in the lwlock.h code previously by
inadvertently adding values after MCPHostsLock and before the line that
reserved each hosts' lock) and furthermore it was causing every process
to run the lock association code, which was pointless.

The MCPFile associated with MCPHosts is now short-lived.  Instead of
keeping the file open all the time, we open on demand and close it
immediately.  This was done mainly to remove the MCPFile member from the
MCPHosts struct, but from a code cleanliness standpoint it also makes
sense.