Send ACK from forwarder to master only when ACKed from slaves
Previously, the forwarder acknowledged messages to the master as soon as
it had them on disk, thus possibly before all slaves had received them.
This meant that if the slaves switch from one forwarder to a different
one, they could be missing a few messages that had not been received and
are now lost because the master pruned them when the forwarder
acknowledged it.  With the new setup, the master will keep those
messages and will be able to send them to the new forwarder, until the
slaves have received and acknowledged them.