Improve calculations for queue pruning, reducing the locking needed
We now keep track of two record numbers for each host: "send next" and
"reading".  When the master process receives a message from the master
to truncate the queue, it needn't disturb the slave; simply make
SendNext jump ahead; the slave process will pick that up in the next
iteration.  The record being read is kept in Reading, and this lets the
pruning process know that it mustn't remove that record.  We previously
kept track of a single record number, requiring an additional lock to
prevent queue truncation, which was very ugly.

MCPHostsGetPruningRecno has been completely rewritten; it now ignores
considerations of queue size.  Those will have to be reintroduced in a
different way later.  This patch also makes some other code be
unreachable, which needs some cleanup.