Simplify queue pruning in forwarder by considering only AckedRecno
We set the queue's AckRecno and FirstRecno to the minimum of all hosts'
AckedRecno, so that the queue prune can remove all records laying before
that value.  This is the intended use of the queue's AckRecno; and as
for FirstRecno, while slightly bogus this is not a problem because
it is unused in the forwarder anyway.

Since the Reading recno is useless, it has been removed.  Per discussion
on the mailing list.