Couple of important fixes:
Check the recno of the table list received from the slave to match the one the forwarder is waiting for. Without this change, the forwarder can clear this flag to early if multiple catalog dumps are sent to slaves.

Don't set the table's requested flag if the table dump is known to be already sent to the slave. If this flag is set - the non-dump transactions for the table are skipped, and this is not what we want in that case.

Several small, but essential one-line fixes.