Fixes for replication of special relations (i.e. pg_largobject).
Previously, we used to include it in a tablelist sent from slave.
That didn't work all the time (due to timing issues, i.e. when
the forwarder had to decide on sending a largeobject transaction to slave
before receiving the complete initial table list). These changes bring
the logic of processing that special case relations to the forwarder itself.