Add necessary flags and data structures to be able to discern between full dump and normal dump transactions (ultimately to skip a full dump).
Per-host flags to tell the host whether to skip the dump, or to accept
the data transactions during full dump (normally the data is skipped for
the slave not in sync).

An additional transaction flag to mark a full dump transaction explicitly
(i.e. to be able to skip a catalog dump that is part of the full dump and
not a stand-alone one).