Remove the code that handles MCP_QUEUE_TRUNCATE message and allow the forwarder to skip sending a full dump to a slave if the slave is in sync.
To differentiate between a full dump transaction and a non-dump one with the same flags (i.e. TABLE_DUMP or CATALOG_DUMP) a FULL_DUMP flag is
appended to such transaction. Previously, the queue was put to sync upon
sending the initial transaction of the full dump. This has been changed to
set the sync state only after the final full dump transaction. Since a slave
with the not-in-sync queue skips data transactions, a separate per-slave flag
(MCP_HOST_FLAG_ACCEPT_DATA_DURING_DUMP) was introduced to handle this.

The generic code for sending queue messages was changed to send MCP_QUEUE_FLAG_EMPTY empty messages instead of silently skipping. This
solves the problem of gaps in the queue, that prevented the next to send/
next to restore record advance on the forwarder/agent slave. Additionally,
the table list hook is now called for the dump transactions, even when
they don't have any tables (since the code that decides whether to send them
should be in the table list hook, and other non-table criteria are involved).

There are a couple of special cases (i.e TABLE_DUMP transactions are
always examined whenever full dump should be skipped or not, pg_largeobject
and repl_slave_roles are specially processed) in the slave table list hook,
which should be re-examined in the future.